fix_move_watcher_and_handler
